# db: mergingIter.SeekPrefixGE stops early if prefix cannot match