--- a/testdata/merging_iter
+++ b/testdata/merging_iter
@@ -502,3 +502,53 @@ next
 ----
 iwoeionch#792,1:792
 .
+
+# Exercise the early stopping behavior for prefix iteration when encountering
+# range deletion tombstones. Keys a, d are not deleted, while the rest are.
+define
+L
+a.SET.10 d.SET.10
+a.SET.10:a10 b.SET.10:b10 c.SET.10:c10 d.SET.10:d10 b.RANGEDEL.12:d
+----
+1:
+  000025:[a#10,SET-d#10,SET]
+
+iter
+first
+next
+next
+----
+a#10,1:a10
+d#10,1:d10
+.
+
+# The seek to c finds d since iteration cannot stop at c as it matches the
+# prefix, and when it steps to d, it finds d is not deleted. Note that
+# mergingIter is an InternalIterator and does not need to guarantee prefix
+# match -- that is job of the higher-level Iterator. So "seek-prefix-ge c" is
+# allowed to return d.
+iter
+seek-prefix-ge a false
+seek-prefix-ge aa true
+seek-prefix-ge b true
+seek-prefix-ge c true
+seek-prefix-ge d true
+----
+a#10,1:a10
+.
+.
+d#10,1:d10
+d#10,1:d10
+
+iter
+seek-prefix-ge a false
+next
+seek-prefix-ge b false
+seek-prefix-ge d true
+next
+----
+a#10,1:a10
+.
+.
+d#10,1:d10
+.